Split OFREP into its own module; factory-only API; align Jackson
