+++ build.sbt
@@ -89,11 +89,13 @@ lazy val commonSettings = Seq(
   )
 ) ++ crossVersionSourceDirs
 
-// Override vulnerable transitive jackson-core (GHSA-72hv-8253-57qq)
-ThisBuild / dependencyOverrides += "com.fasterxml.jackson.core" % "jackson-core" % "2.18.6"
+// Align Jackson with what `ofrep` brings in. GHSA-72hv-8253-57qq (jackson-core <2.18.0) is patched in 2.18+; the OFREP
+// contrib provider pulls 2.21.2, so we override to that version to avoid a split Jackson family (core 2.18 / databind
+// 2.21 leads to NoSuchMethodError at runtime).
+ThisBuild / dependencyOverrides += "com.fasterxml.jackson.core" % "jackson-core" % "2.21.2"
 
 lazy val root = (project in file("."))
-  .aggregate(core, testkit, extras)
+  .aggregate(core, testkit, extras, ofrep)
   .settings(
     name                           := "zio-openfeature",
     publish / skip                 := true,
@@ -114,17 +116,30 @@ lazy val core = (project in file("core"))
     )
   )
 
-// Extras module - built-in providers (HOCON, env vars, caching wrapper, OFREP)
+// Extras module - built-in providers (HOCON, env vars, caching wrapper)
 lazy val extras = (project in file("extras"))
   .dependsOn(core)
   .settings(
     name := "zio-openfeature-extras",
     commonSettings,
     libraryDependencies ++= Seq(
-      "dev.zio"                            %% "zio"       % zioVersion,
-      "dev.zio"                            %% "zio-cache" % "0.2.3",
-      "com.typesafe"                        % "config"    % "1.4.3",
-      "dev.openfeature.contrib.providers"   % "ofrep"     % "0.0.1"
+      "dev.zio"      %% "zio"       % zioVersion,
+      "dev.zio"      %% "zio-cache" % "0.2.3",
+      "com.typesafe"  % "config"    % "1.4.3"
+    )
+  )
+
+// OFREP module - OpenFeature Remote Evaluation Protocol provider.
+// Kept separate from `extras` so callers who only want HOCON/env vars don't pull in the OFREP contrib provider's
+// transitive HTTP-client stack (Jackson, Guava, Commons Validator, SLF4J).
+lazy val ofrep = (project in file("ofrep"))
+  .dependsOn(core)
+  .settings(
+    name := "zio-openfeature-ofrep",
+    commonSettings,
+    libraryDependencies ++= Seq(
+      "dev.zio"                            %% "zio"   % zioVersion,
+      "dev.openfeature.contrib.providers"   % "ofrep" % "0.0.1"
     )
   )
